Fixing Scoped Bug and Updating Catalog Service

+++ Src/Services/CatalogService/Core/E-Commerce.CatalogService.Application/Features/CatalogItems/Dtos/CreateCatalogItemDto.cs
@@ -7,8 +7,8 @@ namespace E_Commerce.CatalogService.Application.Features.CatalogItems.Dtos
         public string Name { get; set; }
         public string Description { get; set; }
         public decimal Price { get; set; }
-        public ushort CatalogBrandId { get; set; }
-        public ushort CatalogTypeId { get; set; }
+        public uint CatalogBrandId { get; set; }
+        public uint CatalogTypeId { get; set; }
         public IList<IFormFile>? ImagesPath { get; set; }
     }
 }
