+++ Src/Services/CatalogService/Core/E-Commerce.CatalogService.Domain/Entity/CatalogItem.cs
@@ -2,7 +2,7 @@
 
 namespace E_Commerce.CatalogService.Domain.Entity
 {
-    public class CatalogItem : BaseEntity<int>
+    public class CatalogItem : BaseEntity<uint>
     {
         public string Name { get; set; }
 
@@ -12,9 +12,9 @@ namespace E_Commerce.CatalogService.Domain.Entity
 
         public CatalogItemFeature CatalogItemFeature { get; set; }
         public CatalogBrand CatalogBrand { get; set; }
-        public ushort CatalogBrandId { get; set; }
+        public uint CatalogBrandId { get; set; }
         public CatalogType CatalogType { get; set; }
-        public ushort CatalogTypeId { get; set; }
+        public uint CatalogTypeId { get; set; }
 
         public ICollection<CatalogItemImage> CatalogItemImages { get; set; }
