+++ Src/Services/CatalogService/Core/E-Commerce.CatalogService.Domain/Entity/CatalogType.cs
@@ -2,7 +2,7 @@
 
 namespace E_Commerce.CatalogService.Domain.Entity
 {
-    public class CatalogType : BaseEntity<ushort>
+    public class CatalogType : BaseEntity<uint>
     {
         public string Type { get; set; }
         public ICollection<CatalogItem> CatalogItems { get; set; }
